+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 0.8.2
+- implement `setReferrerCheck()` method to enable HTTP Referrer check to 
+  avoid CSRF protection for POST/PUT/DELETE request (disabled by default 
+  for now)
+
 ## 0.8.1
 - write all exceptions to `error_log`
 - make it possible to disable plugins by default and enable them on a per